bug 1665692: move PaginationForm to a central location
